requirements.txt
annotated-types==0.7.0
anyio==4.10.0
bcrypt==3.2.2
click==8.1.8
fastapi==0.116.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-auth==2.40.3
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
h11==0.16.0
httptools==0.6.4
idna==3.10
openai
passlib[bcrypt]==1.7.4
pydantic==2.11.7
python-dotenv==1.1.1
python-jose[cryptography]==3.5.0
python-multipart
SQLAlchemy==2.0.43
starlette==0.47.2
typing_extensions==4.14.1
uvicorn==0.35.0
psycopg2-binary